size_t for opcode dataspace size

+++ Top/csound.c
@@ -3311,7 +3311,7 @@ static CS_NOINLINE int32_t opcode_list_new_oentry(CSOUND *csound,
 }
 
 PUBLIC int32_t csoundAppendOpcode(CSOUND *csound,
-                              const char *opname, int32_t dsblksiz, int32_t flags,
+                              const char *opname, size_t dsblksiz, int32_t flags,
                               const char *outypes, const char *intypes,
                               int32_t (*init)(CSOUND *, void *),
                               int32_t (*perf)(CSOUND *, void *),
@@ -3320,8 +3320,8 @@ PUBLIC int32_t csoundAppendOpcode(CSOUND *csound,
   OENTRY  tmpEntry;
   int32_t     err;
   tmpEntry.opname     = (char*) opname;
-  tmpEntry.dsblksiz   = (uint16) dsblksiz;
-  tmpEntry.flags      = (uint16) flags;
+  tmpEntry.dsblksiz   =  dsblksiz;
+  tmpEntry.flags      =  flags;
   tmpEntry.outypes    = (char*) outypes;
   tmpEntry.intypes    = (char*) intypes;
   tmpEntry.init     = init;
